Use Sony's own types in the SDK code
The project's s32/u8 typedefs are a reasonable house style for the
game's own code, but Sony's published headers and reference use u_long,
u_char, u_short, short, long and char, and that is what the SDK code
should read as. 176 files converted across psyq, sdk, libsnd, libspu and
kernel; the game's own directories keep the project typedefs, since we
have no evidence about what Namco used and guessing would be inventing
history.

This is safe because the spelling of a same-width type is free here:
signed int, int, long and signed long compile to byte-identical code,
verified before starting. Only width and signedness matter, and those
are already pinned by byte-exactness at every real access. The one trap
is s8, which maps to signed char rather than char, because the build
passes -funsigned-char. Bare int also had to become long, since gcc
2.6.3 treats them as conflicting in a prototype.

Matching Sony's signatures surfaced a list of disagreements that are
potential misunderstandings rather than style. SsSeqPause takes one
argument in the reference and two in ours, which suggests we have bound
the shared _SsSndPause body under the wrong name. SsSeqOpen takes two
and ours takes one. StSetStream's last two arguments are function
pointers, not a callback and user data pair. DsSearchFile returns a file
pointer, not a long, and the whole event API returns long where we say
void. Seven parameter names were simply wrong and are fixed from the
reference, including SsUtChangePitch's four, which are an old and a new
note-and-fine pair rather than a tone, an unused, a centre and a fine.

The libsnd arguments are deliberately left wide against the reference's
short: the game calls those unprototyped and passes full words, so
narrowing them would change caller codegen.

+++ src/main/PAL/lib/kernel/WaitEvent.c
@@ -1,6 +1,8 @@
+#include <sys/types.h>
+
 #include "common.h"
 
-u32 WaitEvent[4] asm("func_8007B004") __attribute__((section(".text"))) = {
+u_long WaitEvent[4] asm("func_8007B004") __attribute__((section(".text"))) = {
     0x240A00B0,
     0x01400008,
     0x2409000A,
