+++ src/main/PAL/main/cd/GameStepCdPauseRequest.c
@@ -21,10 +21,10 @@ extern s32 g_CdFadeFrames asm("D_8009B1B4");
 extern s32 g_CdTrackEnded asm("D_8019C7BC");
 extern CdlLOC g_CdTrackLoopPoint[] asm("D_8007F5B0");
 
-s32 CdControl(s32 com, void *param, s32 result) asm("func_8006A5A4");
+long CdControl(long com, void *param, long result) asm("func_8006A5A4");
 s32 func_8006A534(s32 arg0, s32 arg1);
 s32 func_8006A554(s32 arg0, void *arg1);
-s32 CdPosToInt_Local(CdlLOC *loc) asm("func_8006AADC");
+long CdPosToInt_Local(CdlLOC *loc) asm("func_8006AADC");
 void GameStepCdTrackRequest(void) asm("func_800432A8");
 void GameStepCdPlayRequest(void) asm("func_80043494");
 void GameStepCdResumeRequest(void) asm("func_800437B8");
