+++ src/main/PAL/main/render/GameSetDrawClipRect.c
@@ -147,7 +147,7 @@ void GameDrawSprite(void *ot, s16 x0, s16 y0, s16 x1, u16 y1, u16 u0, u16 v0, u8
 }
 
 void func_80064EE0(void *arg0);
-void SetSemiTrans(void *arg0, s32 enabled) asm("func_80064E90");
+void SetSemiTrans(void *arg0, long enabled) asm("func_80064E90");
 void *func_80017390(void *ot, void *prim, s32 arg2);
 
 void GameDrawFlatTriangle(void *ot, s16 x0, s16 y0, s16 x1, u16 y1, u16 x2, u16 y2, u8 r, u8 g, u8 b, s32 semiTrans, u32 arg11) asm("func_80046BA0");
@@ -268,8 +268,8 @@ void GameDrawFlatQuad(void *ot, s16 x0, s16 y0, s16 x1, u16 y1, u16 x2, u16 y2, 
 }
 
 void SetPolyFT4(void *prim) asm("func_80064F44");
-void SetShadeTex(void *prim, s32 enabled) asm("func_80064EB8");
-void SetSemiTrans(void *prim, s32 enabled) asm("func_80064E90");
+void SetShadeTex(void *prim, long enabled) asm("func_80064EB8");
+void SetSemiTrans(void *prim, long enabled) asm("func_80064E90");
 
 /*
  * Packs a POLY_FT4 (textured quad) at the scratchpad cursor and links it into
@@ -329,7 +329,7 @@ void GameDrawTexturedQuad(s32 ot, s16 x0, s16 y0, s16 x1,
 }
 
 void func_80064FF8(void *arg0);
-void SetSemiTrans(void *arg0, s32 enabled) asm("func_80064E90");
+void SetSemiTrans(void *arg0, long enabled) asm("func_80064E90");
 void *func_80017390(void *ot, void *prim, s32 arg2);
 
 void GameDrawSolidRect(void *ot, s32 x0, s32 y0, s32 x1, u16 y1, u8 r, u8 g, u8 b, u8 alpha) asm("func_80047024");
@@ -380,7 +380,7 @@ void GameDrawSolidRect(void *ot, s32 x0, s32 y0, s32 x1, u16 y1, u8 r, u8 g, u8 
 }
 
 void func_8006500C(void *arg0);
-void SetSemiTrans(void *arg0, s32 enabled) asm("func_80064E90");
+void SetSemiTrans(void *arg0, long enabled) asm("func_80064E90");
 void *func_80017390(void *ot, void *prim, s32 arg2);
 
 void GameDrawLine(void *ot, s32 x0, s32 y0, s32 x1, u16 y1, u8 r, u8 g, u8 b, u8 alpha) asm("func_8004711C");
@@ -466,7 +466,7 @@ void GameDrawPolyLine3(void *ot, s16 x0, s16 y0, s16 x1, s16 y1, s16 x2, s16 y2,
 }
 
 void func_80065020(void *arg0);
-void SetSemiTrans(void *arg0, s32 enabled) asm("func_80064E90");
+void SetSemiTrans(void *arg0, long enabled) asm("func_80064E90");
 void *func_80017390(void *ot, void *prim, s32 arg2);
 
 void GameDrawGradientLine(void *ot, s32 x0, s32 y0, s32 x1, u16 y1, u8 r0, u8 g0, u8 b0, u8 r1, u8 g1, u8 b1, u8 alpha) asm("func_80047330");
