+++ src/main/PAL/main/sdk/ExitCriticalSection.c
@@ -1,3 +1,5 @@
+#include <sys/types.h>
+
 #include "common.h"
 
 /*
@@ -28,7 +30,7 @@
  */
 INCLUDE_ASM("asm/PAL/main/nonmatchings/main/sdk/ExitCriticalSection", func_80063220);
 
-u32 func_80063230[4] __attribute__((section(".text"))) = {
+u_long func_80063230[4] __attribute__((section(".text"))) = {
     0x240A00B0,
     0x01400008,
     0x24090032,
